npx firebase init function
